+++ NEWS.md
@@ -1,5 +1,6 @@
 # rjtools 0.0.0.9002
 
+* `check_title` suggests correct title format for article title (#22)
 * Designated csl file for journal articles
 * Add RStudio project template: R Journal Article (#19)
 * Add github links (#18)